test_linux: implemented ignition_toggle to change the ignition status
ign_flag was initially designed to call directly the sensor-hal
ignition_on_off function.
In order to have something more similar with how in reality this will
be used, this patch is adding the necessary logic to write and toggle
in the hal_config file the string used to trigger the ignition status.

Signed-off-by: Faiyaz Chaudhury <faiyaz.chaudhury@st.com>
Change-Id: I5d1120a8040ef1d68b43c3b5de4581c847f24e53
Reviewed-on: https://gerrit.st.com/c/asm330lhhx-ford/STMems-Platform1-IIO-SensorHAL/+/265802
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Denis CIOCCA <denis.ciocca@st.com>
Reviewed-by: Mario TESI <mario.tesi@st.com>